+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2026-04-08
+
+- Add note about file-based fallback
+- Ask user for preference on file-based fallback
+
 ## [0.2.1] - 2026-03-23
 
 ### Added